requirements.txt
blinker==1.8.2
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
flask==3.0.3
fonttools==4.57.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
werkzeug==3.0.6
zipp==3.20.2
